Moved sqlalchemy class to sqlalchemy codebase and more generic link classes to types. closes #153
