+++ stac_fastapi/sqlalchemy/stac_fastapi/sqlalchemy/models/database.py
@@ -12,8 +12,8 @@ from sqlalchemy.ext.declarative import declarative_base
 from stac_pydantic.shared import DATETIME_RFC339
 
 from stac_fastapi.sqlalchemy.models import schemas
+from stac_fastapi.sqlalchemy.models.search import QueryableTypes
 from stac_fastapi.types.config import Settings
-from stac_fastapi.types.search import QueryableTypes
 
 BaseModel = declarative_base()
