+++ stac_fastapi/sqlalchemy/stac_fastapi/sqlalchemy/models/decompose.py
@@ -8,13 +8,9 @@ from pydantic import BaseModel
 from pydantic.utils import GetterDict
 from stac_pydantic.shared import DATETIME_RFC339
 
-from stac_fastapi.sqlalchemy.models.links import (
-    CollectionLinks,
-    ItemLinks,
-    filter_links,
-)
 from stac_fastapi.types.config import Settings
 from stac_fastapi.types.errors import DatabaseError
+from stac_fastapi.types.links import CollectionLinks, ItemLinks, filter_links
 
 
 def resolve_links(links: list, base_url: str) -> List[Dict]:
