+++ stac_fastapi/types/stac_fastapi/types/links.py
